requirements.txt
streamlit
crewai
PyPDF2
python-docx
langchain
chromadb
python-dotenv
openai